+++ repository/mock/database.go
@@ -1,6 +1,12 @@
 package mock
 
-import "github.com/library/domain/model"
+import (
+	"fmt"
+	"github.com/library/domain/model"
+	"github.com/library/handler/dto"
+	"strings"
+
+)
 
 var Books = []model.Book{
 	{
@@ -56,4 +62,42 @@ func (bm Db) FindBookById(id int) (book model.Book, location int, found bool) {
 	}
 
 	return book, location, found
+}
+
+
+func CreateBookModelFromBook(b dto.Book) (bm model.Book) {
+
+	Shelf.Id = Shelf.Id + 1
+
+	isbn10 := ""
+	if b.Identifier.ISBN10 != nil {
+		isbn10 = b.Identifier.ISBN10[0]
+	}
+	isbn13 := ""
+	if b.Identifier.ISBN13 != nil {
+		isbn13 = b.Identifier.ISBN13[0]
+	}
+	fmt.Println(Shelf.Id, isbn10, isbn13)
+	CoverId := ""
+	if b.Cover.Url != "" {
+		part1 := strings.Split(b.Cover.Url, "/")[5]
+		part2 := strings.Split(part1, ".")[0]
+		CoverId = strings.Split(part2, "-")[0]
+	}
+	libraryId := ""
+	if b.Identifier.Openlibrary != nil {
+		libraryId = b.Identifier.Openlibrary[0]
+	}
+
+	bookToAdd := model.Book{
+		Id:            Shelf.Id,
+		Title:         b.Title,
+		Author:        b.Author[0].Name,
+		Isbn:          isbn10,
+		Isbn13:        isbn13,
+		OpenLibraryId: libraryId,
+		CoverId:       CoverId,
+		Year:          b.Year,
+	}
+	return bookToAdd
 }
